requirements.txt
pillow
tweepy